Update types for batchRingBufferSize and maxBatchesToVerify
Changed batchRingBufferSize from uint64 to uint32 and maxBatchesToVerify from uint64 to uint8 in the IInbox interface to better reflect their expected value ranges and optimize storage.

+++ packages/protocol/contracts/layer1/based2/IInbox.sol
@@ -205,9 +205,9 @@ interface IInbox {
         /// @notice The chain ID of the network where Taiko contracts are deployed.
         uint64 chainId;
         /// @notice Size of the batch ring buffer, allowing extra space for proposals.
-        uint64 batchRingBufferSize;
+        uint32 batchRingBufferSize;
         /// @notice The maximum number of verifications allowed when a batch is proposed or proved.
-        uint64 maxBatchesToVerify;
+        uint8 maxBatchesToVerify;
         /// @notice The maximum gas limit allowed for a block.
         uint48 livenessBond;
         /// @notice The amount of Taiko token as a proposer's provability bond per batch.
